Transfiguro - Transform or Add New Variables to a Data Frame

transfiguro magically transforms or adds new variables to a data frame, mimicking the functionality of dplyr::mutate(). Use this function to seamlessly modify or extend your data frame with new calculations or transformations, all within the enchanting world of Harry Potter-themed data analysis.

Usage

transfiguro(.data, ...)

Arguments

.data

A data frame or tibble to be transformed.

...

Additional arguments specifying the transformations to be applied. Each argument can be a new column name followed by a formula to calculate its values.

Value

A transformed data frame or tibble with the new or modified variables.

See also

Examples

if (FALSE) { # \dontrun{
library(dplyr)
# Assuming star_data is a data frame with columns: name, house, and year
transfiguro(star_data, age_next_year = year + 1)
} # }